Added in routing. Placed in $CONFIG array. Format is:
$CONFIG['routing'] = array( '#some/pattern/(.*)#' => array( 'action' => 'actionname',
                                                            'extra'  => 1 ) );

The key of the array is a pattern for matching the route.
The value is a mapping array to route bits of the URL into the appropriate places.
This map has 3 keys: action, method, extra. They correspond to the places to send the URL bits.
Their values can either be a number or a string.
The number indicates what match index of the pattern to apply to that bit.
The string indicates a hard-coded mapping to a particular value, regardless of the pattern matches.